// src/policy/domain/enum/governanceObligation.enum.ts
enum GovernanceObligation {
	AUDIT_JUSTIFICATION = "audit_justification",
	AUDIT_SECURITY_EVENT = "audit_security_event",
	IDENTITY_TIMESTAMP_WATERMARK = "identity_timestamp_watermark",
	INTERNAL_TRACEABILITY_WATERMARK = "internal_traceability_watermark",
	REDACT_CC_HEADER = "redact_cc_header",
	REQUIRE_DOWNGRADE_APPROVAL = "require_downgrade_approval",
	REQUIRE_REASON = "require_reason",
}

export { GovernanceObligation };